Use MgmtPermissionsV2 by default
Closes #40192
